+++ commands/que_opinas.py
@@ -1,11 +1,12 @@
 import discord
-import random
+import time
+from utils.choose_random import choose_random
 from utils.cooldowns import add_cooldown, get_cooldown
-from utils.format_text import load_txt_file
+from utils.format_text import load_text, format_text
 
-OPINIONES = load_txt_file("opiniones.txt")
+OPINIONES = load_text("opiniones.txt")
 OPINIONES_COOLDOWN = 30
-FORZAR_OPINIONES_COOLDOWN = 1800 # 30 mins
+FORZAR_OPINIONES_COOLDOWN = 900 # 15 mins
 
 def opiniones_commands(tree, serverList, godUserID):
     @tree.command(name="queopinas", description="Preguntale a Hepatitis B(ot) que opina acerca de lo que se está hablando", guilds=serverList)
@@ -14,7 +15,7 @@ def opiniones_commands(tree, serverList, godUserID):
         if userID != godUserID:
             cooldown = get_cooldown(userID, "queopinas", OPINIONES_COOLDOWN)
             if cooldown > 0:
-                embed = discord.Embed(description=f"# Este comando está en cooldown \n### Esperá **{round(cooldown)} segundos** para volver a usarlo.")
+                embed = discord.Embed(description=f"# Este comando está en cooldown \n### Esperá **{time.strftime('%M:%S', time.gmtime(cooldown))} segundos** para volver a usarlo.")
                 await interaction.response.send_message(embed=embed, ephemeral=True)
                 return
             add_cooldown(userID, "queopinas")
@@ -24,7 +25,8 @@ def opiniones_commands(tree, serverList, godUserID):
             await interaction.response.send_message(embed=embed, ephemeral=True)
             return
     
-        opinión = random.choice(OPINIONES)
+        opinión = choose_random(OPINIONES, canRepeat=True)
+        opinión = await format_text(interaction, opinión, "emoji")
         embed = discord.Embed(description=f"{opinión}")
         await interaction.response.send_message(embed=embed)
 
@@ -34,7 +36,7 @@ def opiniones_commands(tree, serverList, godUserID):
         if userID != godUserID:
             cooldown = get_cooldown(userID, "forzarqueopinas", FORZAR_OPINIONES_COOLDOWN)
             if cooldown > 0:
-                embed = discord.Embed(description=f"# Este comando está en cooldown \n### Esperá **{round(cooldown)} segundos** para volver a usarlo.")
+                embed = discord.Embed(description=f"# Este comando está en cooldown \n### Esperá **{time.strftime('%M:%S', time.gmtime(cooldown))} segundos** para volver a usarlo.")
                 await interaction.response.send_message(embed=embed, ephemeral=True)
                 return
             add_cooldown(userID, "forzarqueopinas")
@@ -51,5 +53,6 @@ def opiniones_commands(tree, serverList, godUserID):
             return
     
         opinión = OPINIONES[realIndex]
+        opinión = await format_text(interaction, opinión, "emoji")
         embed = discord.Embed(description=f"{opinión}")
         await interaction.response.send_message(embed=embed)