+++ utils/format_text.py
@@ -1,8 +1,7 @@
-import discord
 import re
 
-# Cargar texto, separarlo en texto y autores
-def load_txt_file(txtFile: str):
+# Cargar texto, separarlo en texto y autores si hay
+def load_text(txtFile: str):
     frases = []
     with open(txtFile, "r", encoding="utf-8") as file:
         for line in file:
@@ -19,29 +18,32 @@ def load_txt_file(txtFile: str):
                 frases.append(line)
     return frases
 
-# Separar autores , obtener apodo de autores
-async def separate_authors(interaction, authorsRaw):
-    guild = interaction.guild
-
-    authorsAsIDs = re.findall(r'\d+', authorsRaw)
-    IDsAsInt = [int(ids) for ids in authorsAsIDs]
+# Separar texto , obtener usuarios/emojis
+async def format_text(interaction, textRaw: str, type: str):
+    textAsIDs = re.findall(r'\d+', textRaw)
+    IDsAsInt = [int(ids) for ids in textAsIDs]
 
     for id in IDsAsInt:
-        member = None
-        if guild:
-            member = guild.get_member(id)
-        if member:
-            authorsRaw = authorsRaw.replace(f"{id}", f"{member.display_name}", 1)
-            continue
+        if type == "user":
+            guild = interaction.guild
+            if guild:
+                member = guild.get_member(id)
+                if member:
+                    textRaw = textRaw.replace(f"{id}", f"{member.display_name}", 1)
+                    continue
 
-        userAuthor = interaction.client.get_user(id)
-        if userAuthor is None:
-            try:
-                userAuthor = await interaction.client.fetch_user(id)
-            except discord.NotFound:
-                userAuthor = None
-        if userAuthor:
-            authorsRaw = authorsRaw.replace(f"{id}", f"{userAuthor.name}", 1)
-        else:
-            authorsRaw = authorsRaw.replace(f"{id}", "Usuario Desconocido", 1)
-    return authorsRaw
+            userAuthor = interaction.client.get_user(id)
+            if userAuthor:
+                textRaw = textRaw.replace(f"{id}", f"{userAuthor.name}", 1)
+            else:
+                textRaw = textRaw.replace(f"{id}", "Usuario Desconocido", 1)
+        elif type == "emoji":
+            guild = interaction.guild
+            if guild:
+                emoji = guild.get_emoji(id)
+                if emoji:
+                    textRaw = textRaw.replace(f"{id}", f"{emoji}")
+                    continue
+            
+            textRaw = textRaw.replace(f"{id}", "")
+    return textRaw